requirements-dev.txt
pygame-ce>=2.6
pytest>=8.0
ruff>=0.4
